# openapi.yaml
openapi: 3.0.3
info:
  title: RedotPay Clone API
  description: API for user registration, login, OTP, session management, and TON wallet operations.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Authentication
    description: All authentication-related endpoints
  - name: TON
    description: TON wallet and blockchain operations

paths:
  /auth/sign-up:
    post:
      tags: [Authentication]
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                device_name:
                  type: string
                user_agent:
                  type: string
                os:
                  type: string
                ip:
                  type: string
                type:
                  type: string
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              example:
                success: true
                message: "User Created Successfully!"
                data:
                  refresh_token: "..."
                  access_token: "..."
                  user: {}
        '403':
          description: Missing email or password
        '406':
          description: User already exists
        '400':
          description: Something went wrong

  /auth/sign-in:
    post:
      tags: [Authentication]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '404':
          description: User not found
        '405':
          description: Invalid password
        '420':
          description: Device not trusted
        '421':
          description: No active session

  /auth/send-otp:
    post:
      tags: [Authentication]
      summary: Send OTP for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, user_id]
              properties:
                email:
                  type: string
                user_id:
                  type: string
      responses:
        '200':
          description: OTP sent successfully
        '403':
          description: Missing data
        '400':
          description: Failed to send OTP

  /auth/verify-otp:
    post:
      tags: [Authentication]
      summary: Verify the sent OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, user_id, otp]
              properties:
                email:
                  type: string
                user_id:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: OTP verified successfully
        '403':
          description: Missing data
        '400':
          description: OTP expired or verification failed

  /auth/session-device:
    post:
      tags: [Authentication]
      summary: Create or update session for a specific device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, device_name, os, user_id, ip]
              properties:
                email:
                  type: string
                device_name:
                  type: string
                user_agent:
                  type: string
                os:
                  type: string
                user_id:
                  type: string
                ip:
                  type: string
                type:
                  type: string
      responses:
        '200':
          description: Device session set or updated successfully
        '403':
          description: Missing required data

  /auth/check:
    post:
      tags: [Authentication]
      summary: Validate a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Refresh token is valid
        '400':
          description: Refresh token is invalid or expired

  /ton/generate-ton-address:
    post:
      tags: [TON]
      summary: Generate a new TON wallet address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
      responses:
        '200':
          description: TON Address Generated Successfully
          content:
            application/json:
              example:
                success: true
                message: "TON Address Generated Successfully!"
                data:
                  address: "0Q..."
        '403':
          description: Missing user_id
        '400':
          description: Invalid JSON body or failed generation

  /ton/send-ton:
    post:
      tags: [TON]
      summary: Send TON or Jetton (USDT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, amount, receiver, symbol]
              properties:
                user_id:
                  type: string
                amount:
                  type: number
                receiver:
                  type: string
                symbol:
                  type: string
                  enum: [TON, USDT]
      responses:
        '200':
          description: Transaction successful
        '403':
          description: Missing required fields
        '400':
          description: Wallet/Asset not found or Insufficient balance or Transaction failed